Refactor to remove lwc for all related [object]
