+++ force-app/main/default/aura/BackgroundChecksCmp/BackgroundChecksCmp.cmp
@@ -1,8 +1,22 @@
-<aura:component implements="lightning:isUrlAddressable" description="c:PuppyLogs component">
+<aura:component
+    implements="lightning:isUrlAddressable"
+    description="c:PuppyLogs component"
+>
     <aura:attribute name="id" type="String" />
     <!-- Add an "init" handler to handle the first "v.pageReference" value -->
-    <aura:handler name="init" value="{!this}" action="{!c.onPageReferenceChange}"/>
+    <aura:handler
+        name="init"
+        value="{!this}"
+        action="{!c.onPageReferenceChange}"
+    />
     <!-- Add a "change" handler to handle all changes to the "v.pageReference" value -->
-    <aura:handler name="change" value="{!v.pageReference}" action="{!c.onPageReferenceChange}"/>
-    <c:allBackgroundChecksCmp recordId="{!v.id}"></c:allBackgroundChecksCmp>
-</aura:component>	
+    <aura:handler
+        name="change"
+        value="{!v.pageReference}"
+        action="{!c.onPageReferenceChange}"
+    />
+    <c:relatedBackgroundChecksCmp
+        viewAll="true"
+        recordId="{!v.id}"
+    ></c:relatedBackgroundChecksCmp>
+</aura:component>
