+++ force-app/main/default/aura/FilesCmp/FilesCmp.cmp
@@ -1,8 +1,23 @@
-<aura:component implements="lightning:isUrlAddressable" description="c:All Related Files component">
+<aura:component
+    implements="lightning:isUrlAddressable"
+    description="c:All Related Files component"
+>
     <aura:attribute name="id" type="String" />
     <!-- Add an "init" handler to handle the first "v.pageReference" value -->
-    <aura:handler name="init" value="{!this}" action="{!c.onPageReferenceChange}"/>
+    <aura:handler
+        name="init"
+        value="{!this}"
+        action="{!c.onPageReferenceChange}"
+    />
     <!-- Add a "change" handler to handle all changes to the "v.pageReference" value -->
-    <aura:handler name="change" value="{!v.pageReference}" action="{!c.onPageReferenceChange}"/>
-    <c:allFilesCmp recordId="{!v.id}" objectApiName="{!v.object}"></c:allFilesCmp>
-</aura:component>	
+    <aura:handler
+        name="change"
+        value="{!v.pageReference}"
+        action="{!c.onPageReferenceChange}"
+    />
+    <c:relatedFilesCmp
+        viewAll="true"
+        recordId="{!v.id}"
+        objectApiName="{!v.object}"
+    ></c:relatedFilesCmp>
+</aura:component>
