+++ force-app/main/default/aura/LogsCmp/LogsCmp.cmp
@@ -1,8 +1,19 @@
-<aura:component implements="lightning:isUrlAddressable" description="c:All Related Logs component">
+<aura:component
+    implements="lightning:isUrlAddressable"
+    description="c:All Related Logs component"
+>
     <aura:attribute name="id" type="String" />
     <!-- Add an "init" handler to handle the first "v.pageReference" value -->
-    <aura:handler name="init" value="{!this}" action="{!c.onPageReferenceChange}"/>
+    <aura:handler
+        name="init"
+        value="{!this}"
+        action="{!c.onPageReferenceChange}"
+    />
     <!-- Add a "change" handler to handle all changes to the "v.pageReference" value -->
-    <aura:handler name="change" value="{!v.pageReference}" action="{!c.onPageReferenceChange}"/>
-    <c:allLogsCmp recordId="{!v.id}"></c:allLogsCmp>
-</aura:component>	
+    <aura:handler
+        name="change"
+        value="{!v.pageReference}"
+        action="{!c.onPageReferenceChange}"
+    />
+    <c:relatedLogsCmp viewAll="true" recordId="{!v.id}"></c:relatedLogsCmp>
+</aura:component>
