--- force-app/main/default/lwc/allBackgroundChecksCmp/allBackgroundChecksCmp.html
@@ -1,7 +1,0 @@
-<template>
-    <c-all-related-base-cmp record-id={recordId}>
-        <div slot="list">
-            <c-related-background-checks-cmp record-id={recordId} max="10000"></c-related-background-checks-cmp>
-        </div>
-    </c-all-related-base-cmp>
-</template>
