--- force-app/main/default/lwc/allBackgroundChecksCmp/allBackgroundChecksCmp.js
@@ -1,5 +1,0 @@
-import { LightningElement, api } from 'lwc';
-
-export default class AllBackgroundChecksCmp extends LightningElement {
-    @api recordId;
-}
