--- force-app/main/default/lwc/allContinuingEducationUnitsCmp/allContinuingEducationUnitsCmp.html
@@ -1,7 +1,0 @@
-<template>
-    <c-all-related-base-cmp record-id={recordId}>
-        <div slot="list">
-            <c-related-ceus-cmp record-id={recordId} max="10000"></c-related-ceus-cmp>
-        </div>
-    </c-all-related-base-cmp>
-</template>
