--- force-app/main/default/lwc/allContinuingEducationUnitsCmp/allContinuingEducationUnitsCmp.js
@@ -1,5 +1,0 @@
-import { LightningElement, api } from 'lwc';
-
-export default class AllContinuingEducationUnitsCmp extends LightningElement {
-    @api recordId;
-}
