--- force-app/main/default/lwc/allFilesCmp/allFilesCmp.html
@@ -1,7 +1,0 @@
-<template>
-    <c-all-related-base-cmp record-id={recordId}>
-        <div slot="list">
-            <c-related-files-cmp record-id={recordId} max="10000"></c-related-files-cmp>
-        </div>
-    </c-all-related-base-cmp>
-</template>
