--- force-app/main/default/lwc/allFilesCmp/allFilesCmp.js
@@ -1,6 +1,0 @@
-import { LightningElement, api } from 'lwc';
-
-export default class AllFilesCmp extends LightningElement {
-    @api recordId;
-    @api objectApiName;
-}
