--- force-app/main/default/lwc/allLogsCmp/allLogsCmp.html
@@ -1,7 +1,0 @@
-<template>
-    <c-all-related-base-cmp record-id={recordId}>
-        <div slot="list">
-            <c-related-logs-cmp record-id={recordId} max="10000"></c-related-logs-cmp>
-        </div>
-    </c-all-related-base-cmp>
-</template>
