--- force-app/main/default/lwc/allLogsCmp/allLogsCmp.js
@@ -1,5 +1,0 @@
-import { LightningElement, api } from 'lwc';
-
-export default class AllLogsCmp extends LightningElement {
-    @api recordId;
-}
