--- force-app/main/default/lwc/allLogsCmp/allLogsCmp.js-meta.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
-    <apiVersion>55.0</apiVersion>
-    <isExposed>false</isExposed>
-</LightningComponentBundle>
