--- force-app/main/default/lwc/allPatsCmp/__tests__/allPatsCmp.test.js
@@ -1,25 +1,0 @@
-import { createElement } from 'lwc';
-import AllPatsCmp from 'c/allPatsCmp';
-
-describe('c-all-pats-cmp', () => {
-    afterEach(() => {
-        // The jsdom instance is shared across test cases in a single file so reset the DOM
-        while (document.body.firstChild) {
-            document.body.removeChild(document.body.firstChild);
-        }
-    });
-
-    it('TODO: test case generated by CLI command, please fill in test logic', () => {
-        // Arrange
-        const element = createElement('c-all-pats-cmp', {
-            is: AllPatsCmp
-        });
-
-        // Act
-        document.body.appendChild(element);
-
-        // Assert
-        // const div = element.shadowRoot.querySelector('div');
-        expect(1).toBe(1);
-    });
-});
