--- force-app/main/default/lwc/allPatsCmp/allPatsCmp.html
@@ -1,7 +1,0 @@
-<template>
-    <c-all-related-base-cmp record-id={recordId}>
-        <div slot="list">
-            <c-related-pats-cmp record-id={recordId} max="10000"></c-related-pats-cmp>
-        </div>
-    </c-all-related-base-cmp>
-</template>
