--- force-app/main/default/lwc/allPatsCmp/allPatsCmp.js
@@ -1,5 +1,0 @@
-import { LightningElement, api } from 'lwc';
-
-export default class AllPatsCmp extends LightningElement {
-    @api recordId;
-}
