--- force-app/main/default/lwc/allPuppyLogsCmp/allPuppyLogsCmp.js
@@ -1,5 +1,0 @@
-import { LightningElement, api } from 'lwc';
-
-export default class AllPuppyLogsCmp extends LightningElement {
-    @api recordId;
-}
