--- force-app/main/default/lwc/allRelatedBaseCmp/__tests__/allRelatedBaseCmp.test.js
@@ -1,25 +1,0 @@
-import { createElement } from 'lwc';
-import AllRelatedBaseCmp from 'c/allRelatedBaseCmp';
-
-describe('c-all-related-base-cmp', () => {
-    afterEach(() => {
-        // The jsdom instance is shared across test cases in a single file so reset the DOM
-        while (document.body.firstChild) {
-            document.body.removeChild(document.body.firstChild);
-        }
-    });
-
-    it('TODO: test case generated by CLI command, please fill in test logic', () => {
-        // Arrange
-        const element = createElement('c-all-related-base-cmp', {
-            is: AllRelatedBaseCmp
-        });
-
-        // Act
-        document.body.appendChild(element);
-
-        // Assert
-        // const div = element.shadowRoot.querySelector('div');
-        expect(1).toBe(1);
-    });
-});
