--- force-app/main/default/lwc/allRelatedBaseCmp/allRelatedBaseCmp.html
@@ -1,9 +1,0 @@
-<template>
-    <lightning-card>
-        <lightning-breadcrumbs class="slds-p-around_small">
-            <lightning-breadcrumb label={list} onclick={navigateToList}></lightning-breadcrumb>
-            <lightning-breadcrumb label={name} onclick={navigateToObject}></lightning-breadcrumb>
-        </lightning-breadcrumbs>
-    </lightning-card>
-    <slot name="list"></slot>
-</template>
