--- force-app/main/default/lwc/allRelatedBaseCmp/allRelatedBaseCmp.js-meta.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
-    <apiVersion>55.0</apiVersion>
-    <isExposed>true</isExposed>
-    <targets>
-        <target>lightning__AppPage</target>
-    </targets>
-</LightningComponentBundle>
