--- force-app/main/default/lwc/allRelatedFilesCmp/allRelatedFilesCmp.html
@@ -1,9 +1,0 @@
-<template>
-    <lightning-card>
-        <lightning-breadcrumbs class="slds-p-around_small">
-            <lightning-breadcrumb label={list} onclick={navigateToList}></lightning-breadcrumb>
-            <lightning-breadcrumb label={name} onclick={navigateToObject}></lightning-breadcrumb>
-        </lightning-breadcrumbs>
-    </lightning-card>
-    <c-related-files-cmp record-id={recordId} max="10000"></c-related-files-cmp>
-</template>
