+++ force-app/main/default/lwc/relatedBackgroundChecksCmp/relatedBackgroundChecksCmp.html
@@ -4,8 +4,9 @@
         icon-name="standard:password"
         total={total}
         record-id={recordId}
-        field-name={fieldName}
+        object-api-name={objectApiName}
         aura-comp-name="BackgroundChecksCmp"
+        view-all={viewAll}
     >
         <div slot="actions">
             <!-- lightning button for open modal window -->
