+++ force-app/main/default/lwc/relatedCeusCmp/relatedCeusCmp.html
@@ -4,8 +4,9 @@
         icon-name="custom:custom47"
         total={total}
         record-id={recordId}
-        field-name={fieldName}
+        object-api-name={objectApiName}
         aura-comp-name="ContinuingEducationUnitsCmp"
+        view-all={viewAll}
     >
         <div slot="actions">
             <!-- lightning button for open modal window -->
