+++ force-app/main/default/lwc/relatedFilesCmp/relatedFilesCmp.html
@@ -6,6 +6,7 @@
         record-id={recordId}
         field-name={objectApiName}
         aura-comp-name="FilesCmp"
+        view-all={viewAll}
     >
         <div slot="actions">
             <!-- lightning button for upload files modal -->
