+++ force-app/main/default/lwc/relatedLogsCmp/relatedLogsCmp.html
@@ -4,8 +4,9 @@
         icon-name="standard:logging"
         total={total}
         record-id={recordId}
-        field-name={objectApiName}
+        object-api-name={objectApiName}
         aura-comp-name="LogsCmp"
+        view-all={viewAll}
     >
         <div slot="actions">
             <!-- lightning button for open modal window -->
