+++ force-app/main/default/lwc/relatedLogsCmp/relatedLogsCmp.js
@@ -2,13 +2,12 @@ import { LightningElement, api, wire } from "lwc";
 import { ShowToastEvent } from "lightning/platformShowToastEvent";
 import { NavigationMixin } from "lightning/navigation";
 import getRelatedLogs from "@salesforce/apex/LogController.getRelatedLogs";
-import deleteRelatedLog from "@salesforce/apex/LogController.deleteLog";
+import { deleteRecord } from "lightning/uiRecordApi";
 import { refreshApex } from "@salesforce/apex";
 
 import ATLAS_SUPPORT_FIELD from "@salesforce/schema/Log__c.RequestSupportFromAtlas__c";
 import STRESS_FIELD from "@salesforce/schema/Log__c.Stress__c";
 import CLIENT_FIELD from "@salesforce/schema/Log__c.Client__c";
-import CLIENT_REFERENCE from "@salesforce/schema/Log__c.Client__r.Name";
 import SUBMITTER_FIELD from "@salesforce/schema/Log__c.Submitter__c";
 import FACILITATOR_FIELD from "@salesforce/schema/Log__c.Facilitator__c";
 import DATE_FIELD from "@salesforce/schema/Log__c.Date__c";
@@ -20,6 +19,7 @@ import TEAM_SUPPORT_FIELD from "@salesforce/schema/Log__c.RequestSupportFromTeam
 // Import message service features required for publishing and the message channel
 import { publish, MessageContext } from "lightning/messageService";
 import logForm from "@salesforce/messageChannel/logForm__c";
+import StayInTouchSignature from "@salesforce/schema/User.StayInTouchSignature";
 
 const actions = [
     { label: "Edit", name: "edit" },
@@ -63,8 +63,8 @@ const COLS = [
 
 export default class RelatedLogsCmp extends NavigationMixin(LightningElement) {
     @api recordId;
-    @api max = 6;
     @api objectApiName;
+    @api viewAll;
     columns = COLS;
     data = [];
 
@@ -121,9 +121,13 @@ export default class RelatedLogsCmp extends NavigationMixin(LightningElement) {
         }
     }
 
+    get max() {
+        return this.viewAll ? 10000 : 6;
+    }
+
     @wire(getRelatedLogs, { contactId: "$recordId", max: "$max" })
     getLogs(result) {
-        this.wiredLogs = result;
+        this.wiredData = result;
         this.data = null;
         if (result.data && result.data.items) {
             this.data = result.data.items.map((log) => {
@@ -169,7 +173,7 @@ export default class RelatedLogsCmp extends NavigationMixin(LightningElement) {
     }
 
     deleteLog(recordId) {
-        deleteRelatedLog({ recordId: recordId })
+        deleteRecord(recordId)
             .then(() => {
                 this.dispatchEvent(
                     new ShowToastEvent({
@@ -192,19 +196,6 @@ export default class RelatedLogsCmp extends NavigationMixin(LightningElement) {
     }
 
     handleChange() {
-        refreshApex(this.wiredLogs);
-    }
-
-    handleViewAll() {
-        // Navigate to a specific component.
-        this[NavigationMixin.Navigate]({
-            type: "standard__component",
-            attributes: {
-                componentName: "c__LogsCmp"
-            },
-            state: {
-                c__id: this.recordId
-            }
-        });
+        refreshApex(this.wiredData);
     }
 }
