+++ force-app/main/default/lwc/relatedObjectCmp/relatedObjectCmp.html
@@ -1,6 +1,19 @@
 <template>
     <article class="slds-card slds-card_boundary">
         <div class="slds-page-header header">
+            <lightning-breadcrumbs
+                if:true={viewAll}
+                class="slds-p-around_small"
+            >
+                <lightning-breadcrumb
+                    label={list}
+                    onclick={navigateToList}
+                ></lightning-breadcrumb>
+                <lightning-breadcrumb
+                    label={name}
+                    onclick={navigateToObject}
+                ></lightning-breadcrumb>
+            </lightning-breadcrumbs>
             <div class="slds-media slds-media--center slds-has-flexi-truncate">
                 <div class="slds-page-header__col-title">
                     <div class="slds-media slds-media_center">
